Version 3.3.9

+++ poem/Poem/api/internal_views/probes.py
@@ -16,6 +16,7 @@ from rest_framework.response import Response
 from rest_framework.views import APIView
 
 from tenant_schemas.utils import schema_context, get_public_schema_name
+from .utils import error_response
 
 
 class ListProbes(APIView):
@@ -84,166 +85,216 @@ class ListProbes(APIView):
         )
         schemas.remove(get_public_schema_name())
 
-        probe = admin_models.Probe.objects.get(id=request.data['id'])
-        old_name = probe.name
-        try:
-            package_name = request.data['package'].split(' ')[0]
-            package_version = request.data['package'].split(' ')[1][1:-1]
-            package = admin_models.Package.objects.get(
-                name=package_name, version=package_version
-            )
+        if request.tenant.schema_name == get_public_schema_name() and \
+                request.user.is_superuser:
+            try:
+                probe = admin_models.Probe.objects.get(id=request.data['id'])
+                old_name = probe.name
+                package_name = request.data['package'].split(' ')[0]
+                package_version = request.data['package'].split(' ')[1][1:-1]
+                package = admin_models.Package.objects.get(
+                    name=package_name, version=package_version
+                )
 
-        except IndexError:
-            return Response(
-                {'detail': 'You should specify package version.'},
-                status=status.HTTP_400_BAD_REQUEST
-            )
+                old_version = probe.package.version
 
-        except admin_models.Package.DoesNotExist:
-            return Response(
-                {'detail': 'You should choose existing package.'},
-                status=status.HTTP_400_BAD_REQUEST
-            )
+                if package.version != old_version:
+                    probe.name = request.data['name']
+                    probe.package = package
+                    probe.repository = request.data['repository']
+                    probe.docurl = request.data['docurl']
+                    probe.description = request.data['description']
+                    probe.comment = request.data['comment']
+                    probe.user = request.user.username
 
-        old_version = probe.package.version
+                    probe.save()
+                    create_history(probe, probe.user)
 
-        try:
-            if package.version != old_version:
-                probe.name = request.data['name']
-                probe.package = package
-                probe.repository = request.data['repository']
-                probe.docurl = request.data['docurl']
-                probe.description = request.data['description']
-                probe.comment = request.data['comment']
-                probe.user = request.user.username
-
-                probe.save()
-                create_history(probe, probe.user)
-
-                if request.data['update_metrics'] in [True, 'true', 'True']:
-                    metrictemplates = \
-                        admin_models.MetricTemplate.objects.filter(
-                            probekey__name=old_name,
-                            probekey__package__version=old_version
-                        )
-
-                    for metrictemplate in metrictemplates:
-                        metrictemplate.probekey = \
-                            admin_models.ProbeHistory.objects.get(
-                                name=probe.name,
-                                package__version=probe.package.version
+                    if request.data['update_metrics'] in [True, 'true', 'True']:
+                        metrictemplates = \
+                            admin_models.MetricTemplate.objects.filter(
+                                probekey__name=old_name,
+                                probekey__package__version=old_version
                             )
-                        metrictemplate.save()
-                        create_history(metrictemplate, request.user.username)
 
-            else:
-                history = admin_models.ProbeHistory.objects.filter(
-                    name=old_name, package__version=old_version
-                )
-                probekey = history[0]
-                new_data = {
-                            'name': request.data['name'],
-                            'package': package,
-                            'description': request.data['description'],
-                            'comment': request.data['comment'],
-                            'repository': request.data['repository'],
-                            'docurl': request.data['docurl'],
-                            'user': request.user.username
-                        }
-                admin_models.Probe.objects.filter(pk=probe.id).update(
-                    **new_data
-                )
+                        for metrictemplate in metrictemplates:
+                            metrictemplate.probekey = \
+                                admin_models.ProbeHistory.objects.get(
+                                    name=probe.name,
+                                    package__version=probe.package.version
+                                )
+                            metrictemplate.save()
+                            create_history(
+                                metrictemplate, request.user.username
+                            )
 
-                del new_data['user']
-                new_data.update({
-                    'version_comment': update_comment(
-                        admin_models.Probe.objects.get(
-                            id=request.data['id']
-                        )
+                else:
+                    history = admin_models.ProbeHistory.objects.filter(
+                        name=old_name, package__version=old_version
                     )
-                })
-                history.update(**new_data)
+                    probekey = history[0]
+                    new_data = {
+                        'name': request.data['name'],
+                        'package': package,
+                        'description': request.data['description'],
+                        'comment': request.data['comment'],
+                        'repository': request.data['repository'],
+                        'docurl': request.data['docurl'],
+                        'user': request.user.username
+                    }
+                    admin_models.Probe.objects.filter(pk=probe.id).update(
+                        **new_data
+                    )
 
-                # update Metric history in case probekey name has changed:
-                if request.data['name'] != old_name:
-                    for schema in schemas:
-                        with schema_context(schema):
-                            metrics = poem_models.Metric.objects.filter(
-                                probekey=probekey
+                    del new_data['user']
+                    new_data.update({
+                        'version_comment': update_comment(
+                            admin_models.Probe.objects.get(
+                                id=request.data['id']
                             )
+                        )
+                    })
+                    history.update(**new_data)
 
-                            for metric in metrics:
-                                vers = poem_models.TenantHistory.objects.filter(
-                                    object_id=metric.id
+                    # update Metric history in case probekey name has changed:
+                    if request.data['name'] != old_name:
+                        for schema in schemas:
+                            with schema_context(schema):
+                                metrics = poem_models.Metric.objects.filter(
+                                    probekey=probekey
                                 )
 
-                                for ver in vers:
-                                    serialized_data = json.loads(
-                                        ver.serialized_data
-                                    )
+                                for metric in metrics:
+                                    vers = \
+                                        poem_models.TenantHistory.objects.filter(
+                                            object_id=metric.id
+                                        )
 
-                                    serialized_data[0]['fields']['probekey'] = \
-                                        [request.data['name'],
-                                         package.version]
+                                    for ver in vers:
+                                        serialized_data = json.loads(
+                                            ver.serialized_data
+                                        )
 
-                                    ver.serialized_data = json.dumps(
-                                        serialized_data
-                                    )
-                                    ver.save()
+                                        serialized_data[0]['fields'][
+                                            'probekey'
+                                        ] = \
+                                            [request.data['name'],
+                                             package.version]
 
-            return Response(status=status.HTTP_201_CREATED)
+                                        ver.serialized_data = json.dumps(
+                                            serialized_data
+                                        )
+                                        ver.save()
 
-        except IntegrityError:
-            return Response(
-                {'detail': 'Probe with this name already exists.'},
-                status=status.HTTP_400_BAD_REQUEST
+                return Response(status=status.HTTP_201_CREATED)
+
+            except admin_models.Probe.DoesNotExist:
+                return error_response(
+                    status_code=status.HTTP_404_NOT_FOUND,
+                    detail='Probe does not exist.'
+                )
+
+            except IndexError:
+                return error_response(
+                    status_code=status.HTTP_400_BAD_REQUEST,
+                    detail='Package version should be specified.'
+                )
+
+            except admin_models.Package.DoesNotExist:
+                return error_response(
+                    status_code=status.HTTP_404_NOT_FOUND,
+                    detail='Package does not exist.'
+                )
+
+            except IntegrityError:
+                return error_response(
+                    status_code=status.HTTP_400_BAD_REQUEST,
+                    detail='Probe with this name already exists.'
+                )
+
+            except KeyError as e:
+                return error_response(
+                    status_code=status.HTTP_400_BAD_REQUEST,
+                    detail='Missing data key: {}'.format(e.args[0])
+                )
+
+        else:
+            return error_response(
+                status_code=status.HTTP_401_UNAUTHORIZED,
+                detail='You do not have permission to change probes.'
             )
 
     def post(self, request):
-        try:
-            package_name = request.data['package'].split(' ')[0]
-            package_version = request.data['package'].split(' ')[1][1:-1]
-            probe = admin_models.Probe.objects.create(
-                name=request.data['name'],
-                package=admin_models.Package.objects.get(
-                    name=package_name, version=package_version
-                ),
-                repository=request.data['repository'],
-                docurl=request.data['docurl'],
-                description=request.data['description'],
-                comment=request.data['comment'],
-                user=request.user.username,
-                datetime=datetime.datetime.now()
-            )
-
-            if request.data['cloned_from']:
-                clone = admin_models.Probe.objects.get(
-                    id=request.data['cloned_from']
+        if request.tenant.schema_name == get_public_schema_name() and \
+                request.user.is_superuser:
+            try:
+                package_name = request.data['package'].split(' ')[0]
+                package_version = request.data['package'].split(' ')[1][1:-1]
+                probe = admin_models.Probe(
+                    name=request.data['name'],
+                    package=admin_models.Package.objects.get(
+                        name=package_name, version=package_version
+                    ),
+                    repository=request.data['repository'],
+                    docurl=request.data['docurl'],
+                    description=request.data['description'],
+                    comment=request.data['comment'],
+                    user=request.user.username,
+                    datetime=datetime.datetime.now()
                 )
-                comment = 'Derived from {} ({}).'.format(
-                    clone.name, clone.package.version
+
+                if 'cloned_from' in request.data and \
+                        request.data['cloned_from']:
+                    try:
+                        clone = admin_models.Probe.objects.get(
+                            id=request.data['cloned_from']
+                        )
+                        comment = 'Derived from {} ({}).'.format(
+                            clone.name, clone.package.version
+                        )
+                        probe.save()
+                        create_history(probe, probe.user, comment=comment)
+
+                    except admin_models.Probe.DoesNotExist:
+                        return error_response(
+                            status_code=status.HTTP_404_NOT_FOUND,
+                            detail='Probe from which to clone does not exist.'
+                        )
+
+                else:
+                    probe.save()
+                    create_history(probe, probe.user)
+
+                return Response(status=status.HTTP_201_CREATED)
+
+            except IntegrityError:
+                return error_response(
+                    status_code=status.HTTP_400_BAD_REQUEST,
+                    detail='Probe with this name already exists.'
                 )
-                create_history(probe, probe.user, comment=comment)
 
-            else:
-                create_history(probe, probe.user)
+            except admin_models.Package.DoesNotExist:
+                return error_response(
+                    status_code=status.HTTP_400_BAD_REQUEST,
+                    detail='Package does not exist.'
+                )
 
-            return Response(status=status.HTTP_201_CREATED)
+            except IndexError:
+                return error_response(
+                    status_code=status.HTTP_400_BAD_REQUEST,
+                    detail='Package version should be specified.'
+                )
 
-        except IntegrityError:
-            return Response({'detail': 'Probe with this name already exists.'},
-                            status=status.HTTP_400_BAD_REQUEST)
+            except KeyError as e:
+                return error_response(
+                    status_code=status.HTTP_400_BAD_REQUEST,
+                    detail='Missing data key: {}'.format(e.args[0])
+                )
 
-        except admin_models.Package.DoesNotExist:
-            return Response(
-                {'detail': 'You should choose existing package.'},
-                status=status.HTTP_400_BAD_REQUEST
-            )
-
-        except IndexError:
-            return Response(
-                {'detail': 'You should specify package version.'},
-                status=status.HTTP_400_BAD_REQUEST
+        else:
+            return error_response(
+                status_code=status.HTTP_401_UNAUTHORIZED,
+                detail='You do not have permission to add probes.'
             )
 
     def delete(self, request, name=None):
@@ -251,36 +302,48 @@ class ListProbes(APIView):
             Tenant.objects.all().values_list('schema_name', flat=True)
         )
         schemas.remove(get_public_schema_name())
-        if name:
-            try:
-                probe = admin_models.Probe.objects.get(name=name)
-                mt = admin_models.MetricTemplate.objects.filter(
-                    probekey=admin_models.ProbeHistory.objects.get(
-                        name=probe.name, package__version=probe.package.version
+        if request.tenant.schema_name == get_public_schema_name() and \
+                request.user.is_superuser:
+            if name:
+                try:
+                    probe = admin_models.Probe.objects.get(name=name)
+                    mt = admin_models.MetricTemplate.objects.filter(
+                        probekey=admin_models.ProbeHistory.objects.get(
+                            name=probe.name,
+                            package__version=probe.package.version
+                        )
                     )
-                )
-                if len(mt) == 0:
-                    for schema in schemas:
-                        # need to iterate through schemas because of foreign
-                        # key in Metric model
-                        with schema_context(schema):
-                            admin_models.ProbeHistory.objects.filter(
-                                object_id=probe
-                            ).delete()
-                    probe.delete()
-                    return Response(status=status.HTTP_204_NO_CONTENT)
-                else:
-                    return Response(
-                        {'detail': 'You cannot delete Probe that is associated '
-                                   'to metric templates!'},
-                        status=status.HTTP_400_BAD_REQUEST
-                    )
+                    if len(mt) == 0:
+                        for schema in schemas:
+                            # need to iterate through schemas because of foreign
+                            # key in Metric model
+                            with schema_context(schema):
+                                admin_models.ProbeHistory.objects.filter(
+                                    object_id=probe
+                                ).delete()
+                        probe.delete()
+                        return Response(status=status.HTTP_204_NO_CONTENT)
+                    else:
+                        return error_response(
+                            status_code=status.HTTP_400_BAD_REQUEST,
+                            detail='You cannot delete probe that is associated '
+                                   'to metric templates.'
+                        )
 
-            except admin_models.Probe.DoesNotExist:
-                raise NotFound(status=404, detail='Probe not found')
+                except admin_models.Probe.DoesNotExist:
+                    raise NotFound(status=404, detail='Probe does not exist.')
+
+            else:
+                return error_response(
+                    status_code=status.HTTP_400_BAD_REQUEST,
+                    detail='Probe name not specified.'
+                )
 
         else:
-            return Response(status=status.HTTP_400_BAD_REQUEST)
+            return error_response(
+                status_code=status.HTTP_401_UNAUTHORIZED,
+                detail='You do not have permission to delete probes.'
+            )
 
 
 class ListPublicProbes(ListProbes):
